Added middleware and started consuming backend APIs with React

+++ src/Services/Subscription/Milo.Subscription.Application/Features/AccountInfos/Handlers/RemoveAccountInfoCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Milo.Subscription.Application.Exceptions;
 using Milo.Subscription.Application.Features.AccountInfos.Commands;
 using Milo.Subscription.Application.Interfaces.Repositories;
 using Milo.Subscription.Application.Interfaces.Services;
@@ -19,9 +20,15 @@ namespace Milo.Subscription.Application.Features.AccountInfos.Handlers
         public async Task Handle(RemoveAccountInfoCommand request, CancellationToken cancellationToken)
         {
             var value = await _repository.GetByIdAsync(request.AccountInfoId);
+
+            if (value is null)
+            {
+                throw new NotFoundException("Hesap bilgisi bulunamadı.");
+            }
+
             if (value.UserId != _currentUser.GetUserId())
             {
-                throw new Exception("Bu hesap bilgisine erişim yetkiniz yok.");
+                throw new ForbiddenException("Bu hesap bilgisine erişim yetkiniz yok.");
             }
 
             await _repository.DeleteAsync(request.AccountInfoId);
