+++ src/Services/Subscription/Milo.Subscription.Application/Features/UserSubscriptions/Handlers/RemoveUserSubscriptionCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Milo.Subscription.Application.Exceptions;
 using Milo.Subscription.Application.Features.UserSubscriptions.Commands;
 using Milo.Subscription.Application.Interfaces.Repositories;
 using Milo.Subscription.Application.Interfaces.Services;
@@ -19,9 +20,15 @@ namespace Milo.Subscription.Application.Features.UserSubscriptions.Handlers
         public async Task Handle(RemoveUserSubscriptionCommand request, CancellationToken cancellationToken)
         {
             var value = await _repository.GetByIdAsync(request.UserSubscriptionId);
+
+            if (value is null)
+            {
+                throw new NotFoundException("Abonelik bulunamadı.");
+            }
+
             if (value.UserId != _currentUser.GetUserId())
             {
-                throw new Exception("Bu aboneliği silme yetkiniz yok.");
+                throw new ForbiddenException("Bu aboneliği silme yetkiniz yok.");
             }
 
             await _repository.DeleteAsync(request.UserSubscriptionId);
